[lldb/Utility] Make Scalar only use llvm's arbitrary precision types (WIP)
This patch removes Scalar's fundamental type constructors and assignment
operator overloadings to rely mostly on llvm::APInt and llvm::APFloat.

For the sake of convenience, Scalar still have constructor and assignment
for addresses (lldb::addr_t) and big float (long double).

At the moment, the refactor is done, but the testsuite shows a lot of
failures.

WORK IN PROGRESS.NOT READY FOR REVIEW.

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>