[lldb/Plugins] Introduce Scripted Platform Plugin
This patch introduces Scripted Platform, a new platform plugin that can
be customized with a python script.

For now this can list processes described in the python script file but
eventually, it will be used to spawn scripted processes and act as an
interface between them.

This was already reviewed in https://reviews.llvm.org/D139252

Signed-off-by: Med Ismail Bennani <ismail@bennani.ma>