[lldb/tool] Add SBGenerator clang tool
This patch introduces a new lldb SBGenerator clang tool.

The main goal is to iterate over the lldb SBAPI headers and collect
metadata that will be used to dynamically augment the SBAPI by adding
new classes or new methods to existing classes.

This first usage of this will be to generate the SBOptional class since
we need to make an constructor overload and getter for every SB type.

Signed-off-by: Med Ismail Bennani <ismail@bennani.ma>