[lldb/Breakpoint] Add "inject-condition" flag for Jitted Conditional Breakpoints
Summary:
This patch introduces Fast Conditional Breakpoint (FCB) by adding
a flag to evaluate the condition in-process instead of doing
round-trip between LLDB and the inferior.

The first step is to add support to the CommandObjectBreakpoint
class to recognize the `-I | --inject-condition` options.
This will set the "InjectCondition" BreakpointOptions, but also propagate
it in the Breakpoint, BreakpointLocation and EvaluateExpressionOptions
class, by adding a "inject_condition" member and a getters.

The BreakpointOptions description has been updated to reflect if
the condition expression is JIT-ed. The Fast Conditional Breakpoint
have now a "(FAST)" tag next to them when being listed. The option is
also taken into account when copying the BreakpointOption from one
breakpoint to another.

This patch adds a python test for Fast Conditional Breakpoints.
The test creates a conditional breakpoint on the main function.
It checks if it was correctly created, and if the condition is actually
is set to be jitted.

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>

Subscribers: lldb-commits

Tags: #lldb

Differential Revision: https://reviews.llvm.org/D66248

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>

[lldb/Breakpoint] Add "BreakpointInjectedSite" and FCB Trampoline

Summary:
BreakpointInjectedSite is a sub-class of the BreakpointSite class.
Its main purpose is to handle Fast Conditional Breakpoint setup.

In the case of Fast Conditional Breakpoints, the process will copy
enough instructions to a new memory space and branch to it.

Then it creates a BreakpointInjectedSite that will build the condition
expression, by fetching the condition from each BreakpointLocation
and depending on the architecture and the source language, inject
a trap instruction that will be caught by LLDB at runtime. To do so, the
trap address is resolved and added to LLDB’s BreakpointSiteList.
The BreakpointInjectedSite also initialize the `$__lldb_arg` structure
for the conditional expression using Debug Info.
Right now, it only works with un-optimized (“-O0”) code, on x86_64.

After creating the Fast Conditional Breakpoint, the process will set up the
FCB Trampoline for the current ABI. Since the JIT-ed code might modify the
registers state, we need to save the current context, run the JIT-ed code
and the copied instruction, then restore the context and branch back.

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>

Reviewers: jdoerfert

Subscribers: mgorny, aprantl, lldb-commits

Tags: #lldb

Differential Revision: https://reviews.llvm.org/D66249

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>

[lldb/ObjectFile] Add Trampoline ObjectFile and UnwindPlan support for FCB

Summary:
When hitting a Fast Conditional Breakpoint, LLDB will stop at the JIT-ed
expression level instead of stopping where the user sets the breakpoint.
In order to unwind to user code, LLDB needs symbolic information
for the trampoline taking into consideration that the trampoline altered the
memory layout by spilling registers to the stack.

To solve that, this patches introduces a Trampoline ObjectFile and UnwindPlan,
which creates a module for the Trampoline by setting up the section and the
symbol table. LLDB inserts a FuncUnwinder in the module UnwindTable and
creates an UnwindPlan pointing to the BreakpointLocation return address.
Then it sets the FuncUnwinder to the newly created Trampoline UnwindPlan.
Finally, the newly created module is appended to the target image list, which
allows LLDB to move back and forth between the JIT-ed and the user code.

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>

Subscribers: mgorny, lldb-commits

Tags: #lldb

Differential Revision: https://reviews.llvm.org/D66250

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>

[lldb/Breakpoint] Fix Fast Conditional Breakpoints errors

There was 2 error when trying to use Fast Conditional Breakpoints with
the recent codebase:

1. When gathering the local variables in the Jitted expression with the
   ClangExpressionDeclMap, fetching the expression variable list had 0
   entry making the following stages fail. This was fixed by changing
   the logic to ask right away for an ExpressionVariable at a specific
   index.
2. Since the original implementation, the Utility Function emitting
   process changed. Now, creating the UtilityFunction also installs it
   into the process, and since I was originally trying to do it
   manually, I was getting a error.

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>

[lldb/Process] Check Process ABI pointer before dereferencing it.

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>

[lldb/Process] Fix FCB fallback infinite loop

[lldb] Add support for Aarch64 to FCB

[lldb] Update logging for Fast Conditional Breakpoints

[lldb] Update ObjectFileTrampoline to conform to base class refactor

[lldb/ABI] Start porting FCB trampoline jump to arm64

[lldb/breakpoints] Make FCB work on Aarch64 (WIP)

Signed-off-by: Med Ismail Bennani <medismail.bennani@gmail.com>

[lldb] Make FCB build

Signed-off-by: Med Ismail Bennani <ismail@bennani.ma>