[lldb/API] Introduce SBOptional class
This class acts as a wrapper around other SB instances.

It makes use of `std::any` to store any kind of object and keeps that of
the type it was initialized with by storing a enum value for each type.

I was thinking using std::variant but ran into some issues when some
move-only SB types (i.e. SBStream).

Signed-off-by: Med Ismail Bennani <ismail@bennani.ma>